requirements.txt
anthropic>=0.45.0
gspread>=5.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
xlsxwriter>=3.0.0
# 수집 자동화 (run_collect_sns.py)
requests>=2.28.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
playwright-stealth>=2.0.0
DrissionPage>=4.0.0
# 웹 API 서버
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
